Fixed Reminders. Wasn't working.
Context was always set to "first window" of hexchat,
which was always my "Remi_Scarlet" window. However, I also had
a check so that it would only run on the context of nick
"amatsukaze" or "amatsukaze|bot" so it ended up never running.
Fixed by adding a hook which would update upon finding a context with
the right nick on it and using that hook with correct context to
run the reminders.